Fix PrismaClient initialization in web build by setting up @mindful-mosaic/db package
- Created packages/db/index.js to re-export PrismaClient
- Updated schema.prisma output to ./client in db package
- Added @mindful-mosaic/db workspace dep to apps/web
- Updated imports in lib/auth.ts and routines/route.ts to use @mindful-mosaic/db
- Added build/generate scripts to packages/db/package.json
- Fixed setupTests.ts MockResponse and fetch polyfill
- Formatted files with Prettier
- Cleaned up auth route handler
