+++ apps/web/lib/auth.ts
@@ -1,0 +1,32 @@
+import { NextAuthOptions, Session } from 'next-auth';
+import type { JWT } from 'next-auth/jwt';
+import GoogleProvider from 'next-auth/providers/google';
+import { PrismaAdapter } from '@auth/prisma-adapter';
+import { PrismaClient } from '@mindful-mosaic/db';
+
+let prisma: PrismaClient;
+
+const getPrisma = () => {
+  if (!prisma) {
+    prisma = new PrismaClient();
+  }
+  return prisma;
+};
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(getPrisma()),
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  callbacks: {
+    session({ session, token }: { session: Session; token: JWT }): Session {
+      if (session.user && token.sub) {
+        session.user.id = token.sub as string;
+      }
+      return session;
+    },
+  },
+};
