+++ apps/web/package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "react": "^18",
     "react-dom": "^18",
-    "next": "14.2.0"
+    "next": "14.2.0",
+    "@mindful-mosaic/db": "workspace:*"
   },
   "devDependencies": {
     "typescript": "^5",
