+++ apps/web/src/app/api/auth/[...nextauth]/route.ts
@@ -1,34 +1,5 @@
-import NextAuth, { NextAuthOptions, Session, JWT } from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-import { PrismaAdapter } from '@auth/prisma-adapter';
-import { PrismaClient } from '@prisma/client';
-
-let prisma: PrismaClient;
-
-const getPrisma = () => {
-  if (!prisma) {
-    prisma = new PrismaClient();
-  }
-  return prisma;
-};
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(getPrisma()),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-  ],
-  callbacks: {
-    session({ session, token }: { session: Session; token: JWT }): Session {
-      if (session.user && token.sub) {
-        session.user.id = token.sub as string;
-      }
-      return session;
-    },
-  },
-};
+import NextAuth from 'next-auth';
+import { authOptions } from '../../../../../lib/auth';
 
 const handler = NextAuth(authOptions);
