+++ packages/db/index.js
@@ -1,0 +1,3 @@
+const { PrismaClient } = require('./client');
+
+module.exports = { PrismaClient };
