+++ packages/db/prisma/schema.prisma
@@ -3,7 +3,7 @@
 
 generator client {
   provider = "prisma-client-js"
-  output   = "../../../node_modules/.prisma/client"
+  output   = "../client"
 }
 
 datasource db {
